Fixed Flask app attempting to run before the MySQL server was ready.
Issue was due to a false positive caused by the image starting a temporary server. "--protocol tcp" should prevent the false positive as the initialization runs with --skip-networking (no TCP), so this will have it perform the ping on TCP specifically to check if it's actually ready or not.

Source of fix: https://stackoverflow.com/questions/70485527/mysql-docker-check-if-mysql-server-is-up-and-running